+++ priv/repo/migrations/20211104125950_move_goal_and_minimum_from_rule_to_implementation.exs
@@ -3,8 +3,8 @@ defmodule TdDd.Repo.Migrations.MoveGoalAndMinimumFromRuleToImplementation do
 
   def change do
     alter table("rule_implementations") do
-      add(:goal, :integer)
-      add(:minimum, :integer)
+      add(:goal, :float)
+      add(:minimum, :float)
       add(:result_type, :string, default: "percentage")
     end
